requirements.txt
gradio>=4.0.0
python-docx>=1.1.0
langchain>=0.1.0
openai>=1.0.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
lxml>=4.9.0
requests>=2.31.0
